lang/python/code/drfdemo/requirements.txt
asgiref==3.10.0
Django==5.2.7
djangorestframework==3.16.1
itsdangerous==2.2.0
PyJWT==2.10.1
PyMySQL==1.1.2
sqlparse==0.5.3
typing_extensions==4.15.0
